Chore: Split AttributeFilter files
